Stop hard-coding flux-system
There were hard-coded constants in api/v1alpha1 and ui/lib/types that
both provided a `flux-system` constant. These were used in many places
as The Namespace to use - in particular, they were used for locating
the OIDC/username-password secrets.

There were also several places in our helm charts that hard-coded the
flux-system namespace - both for creating the secrets mentioned above,
and for setting permissions to the flux-system namespace.

This thus means that if gitops and flux run in different namespaces,
and you turn off enableClusterRole, then you will no longer see the
flux runtime - in exchange you do get to see the namespace it's
installed in. Frankly, this is what I'd expect.

This fixes #2415.